Struct rump::client::Client

pub struct Client {
    url: String,
    realm: String,
}

A Client defines methods and options for building a Session with a WAMP Router

Examples

To obtain a Session from a client with default settings perform the following:

let session = Client::new(ADDR_OF_ROUTER, REALM_NAME).connect().unwrap();

Fields

url

realm

Methods

impl Client

fn new(url: &str, realm: &str) -> Self

fn connect(&self) -> WampResult<Session<WebSocket>>