Struct rump::client::Session

pub struct Session<S: WampSender> {
    sender: S,
    subscriptions: Arc<Mutex<(HashMap<String, u64>, HashMap<u64, Vec<Box<Fn(EventPublish<(), ()>) + Send>>>)>>,
    state: SessionState,
}

A Session represents a valid WAMP Session with a Router. You can obtain a Session from a Client

Fields

sender

Outgoing socket connection to WAMP Router

subscriptions

state

Methods

impl<S: WampSender> Session<S>

fn join(&self, realm: String) -> WampResult<()>

Connects to a WAMP Router in a realm without authentication

fn publish<A, K>(&self, topic: &str, args: Vec<WampEncodable<A>>, kwargs: K) where A: Encodable, K: Encodable

Publish an event to the realm You can publish 0 or more positional arguments that are of type WampEncodable and/or a struct representings Key-Value pairs.

# Examples To send an empty publish event session.publish::<(), WampEncodable<()>>("com.example.topic", Vec::new(), WampEncodable::None);

To send the positions arguments (42, "foo") and key word arguments: Note: Your custom type must have the (Encodable)[https://doc.rust-lang.org/rustc-serialize/rustc_serialize/trait.Encodable.html] trait. ```

[derive(RustcEncodable)]

struct CustomKwargs { key1: String key2: u32 }

session.publish::<(), CustomKwargs>("com.example.topic, vec![WampEncodable::i32(42), WampEncodable::String("foo".to_string())], CustomKwargs {key1: "hello".to_string(), key2: 19}); ```

fn subscribe<F>(&self, topic: &str, callback: F) where F: 'static + Send + Fn(EventPublish<(), ()>)