Module rump::message

Reexports

extern crate rustc_serialize;

extern crate rand;

use std::prelude::v1::*;

use rand::Rng;

use rustc_serialize::{Encodable, Encoder, Decodable, Decoder};

use options::{Options, Details};

use transport::Serializer;

use std::result;

use std::collections::HashMap;

use WampError;

use WampResult;

Structs

EventJoin

EventPublish

EventSubscribe

Payload
A struct representing the payload that's received from a WAMP event.

Enums

MessageType
All WAMP events and message types and their numeric counterparts

WampEvent

WampType
All types that can be sent to/from a WAMP Router Used to publish hetereogenous arguments

Functions

new_event_id
Generates a new event_id to track messages sent to and from the WAMP Router