Module rump::transport

Reexports

extern crate rustc_serialize;

extern crate websocket;

use std::prelude::v1::*;

use std::{thread};

use std::sync::mpsc;

use websocket::header::{WebSocketProtocol};

use websocket::client::request::Url;

use websocket::{message, Message, Sender, Receiver};

use rustc_serialize::{Encodable, Encoder, Decodable};

use rustc_serialize::json;

use super::WampResult;

use super::WampError;

Structs

Serializer
Describes the underling Serialization types used

WebSocket
The default socket type used for establishing a WAMP session.

Enums

SerializerType
A type enumerating all possible serialization engines

SocketType
A type enumerating all the possible underlying socket implementations.

Traits

WampConnector
A WampConnector defines methods for a custom socket type to connect to another endpoint and to receive message on the socket

WampSender
A WampSender defines methods for the custom socket type to send over the endpoint